OTA: corrige les timeouts a debordement non signe (maitre + esclave), fw 1.3.7
Les horodatages (_lastActivityMs, _lastSendMs, _serialWaitSince,
_rebootStartMs) sont ecrits avec un millis() frais depuis le traitement de
la boite aux lettres ou la tache Wi-Fi, donc potentiellement POSTERIEURS au
nowMs capture en debut de loop(). En arithmetique non signee, la difference
negative deborde (~4 milliards) et tous les timeouts sautaient
instantanement de facon probabiliste (il suffit qu'un tick de ms tombe entre
le debut de loop() et l'horodatage) : sessions esclave abandonnees 5 ms
apres leur demarrage (observe au moniteur serie), retransmissions et faux
"timeout" cote maitre, blocages a des chunks aleatoires (21, 39, 60, 69...).

Corrige par des comparaisons signees (int32_t)(now - ts) > timeout, l'idiome
standard tolerant au desordre et au wrap. Ajoute au passage des logs serie
de diagnostic sur l'esclave (raison exacte de chaque refus de session).

+++ src/ota_master.cpp
@@ -208,11 +208,18 @@ void OtaMaster::update(uint32_t nowMs) {
 
     {
         CriticalGuard guard(_mux);
+        // Comparaisons SIGNEES obligatoires : _lastSendMs/_serialWaitSince/
+        // _rebootStartMs sont horodates avec un millis() frais depuis onAck()
+        // (tache Wi-Fi) ou sendFrame() (appele par Console.update() PLUS TOT
+        // dans la meme iteration de loop()) — ils peuvent donc etre POSTERIEURS
+        // au nowMs capture en debut de loop(). En non signe, la difference
+        // negative deborde et declenche retransmissions et faux "timeout"
+        // instantanes (meme bug que cote esclave, voir ota_slave.cpp).
         switch (_state) {
         case OM_AWAIT_START_ACK:
         case OM_AWAIT_CHUNK_ACK:
         case OM_AWAIT_END_ACK:
-            if (nowMs - _lastSendMs > OTA_ACK_TIMEOUT_MS) {
+            if ((int32_t)(nowMs - _lastSendMs) > (int32_t)OTA_ACK_TIMEOUT_MS) {
                 if (++_retryCount > OTA_MAX_RETRIES) {
                     fail("timeout");
                     needAbort = _pendingAbortNeeded;
@@ -224,7 +231,7 @@ void OtaMaster::update(uint32_t nowMs) {
             break;
 
         case OM_AWAIT_SERIAL_CHUNK:
-            if (nowMs - _serialWaitSince > OTA_SERIAL_IDLE_TIMEOUT_MS) {
+            if ((int32_t)(nowMs - _serialWaitSince) > (int32_t)OTA_SERIAL_IDLE_TIMEOUT_MS) {
                 fail("console injoignable");
                 needAbort = _pendingAbortNeeded;
                 if (needAbort) { abortTarget = _pendingAbortTarget; abortSession = _pendingAbortSession; _pendingAbortNeeded = false; }
@@ -252,7 +259,7 @@ void OtaMaster::update(uint32_t nowMs) {
                 break;
             }
             (void)found;
-            if (_state == OM_AWAIT_REBOOT && nowMs - _rebootStartMs > OTA_REBOOT_WAIT_MS) {
+            if (_state == OM_AWAIT_REBOOT && (int32_t)(nowMs - _rebootStartMs) > (int32_t)OTA_REBOOT_WAIT_MS) {
                 _pending = Event{};
                 _pending.type = EV_RESULT;
                 _pending.target = _target;
